pyproject.toml
[project]
name = "ensemble-phase-2-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cohere>=5.20.2",
    "langchain>=1.2.7",
    "langchain-cohere>=0.5.0",
    "langgraph>=1.0.7",
    "mlflow==3.0.1",
    "python-dotenv>=1.2.1",
]
